mods to accomodate the new state process

+++ def_stats.cpp
@@ -77,10 +77,10 @@ void def_stats::wtrans_iterate(int t_hor_in) {
 					w2_l = (*snodes1).w_t2_state_low[i_t_hor][i_s1p][i_s2p][i_w1p];    // load in wealth in next period given state
 					w2_h = (*snodes1).w_t2_state_high[i_t_hor][i_s1p][i_s2p][i_w1p];    // (low and high wealth realizations)
 
-					i_w2l = round((w2_l - w_min) / (w_max - w_min)*(w_n-1));   // round wealth realization to closest wealth index
+					i_w2l = (int) round((w2_l - w_min) / (w_max - w_min)*(w_n-1));   // round wealth realization to closest wealth index
 					i_w2l = min(max(i_w2l, 0), w_n - 1);               // TODO: double check round fn
 
-					i_w2h = round((w2_h - w_min) / (w_max - w_min)*(w_n-1));
+					i_w2h = (int) round((w2_h - w_min) / (w_max - w_min)*(w_n-1));
 					i_w2h = min(max(i_w2h, 0), w_n - 1);
 
 					wdist2[i_w2l] = wdist2[i_w2l] + 0.5*sdist[i_s1p] * gammap[i_s1p][i_s2p] * wdist[i_w1p];  // add outcome probability mass
